api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

pydantic==2.6.4
pydantic[email]==2.6.4
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.9
python-dotenv==1.0.1